# Fix: estraverse/escope to work with unknowns (fixes #5476)