--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -11,8 +11,7 @@
 //------------------------------------------------------------------------------
 
 var lodash = require("lodash"),
-    espree = require("espree"),
-    estraverse = require("estraverse"),
+    Traverser = require("./util/traverser"),
     escope = require("escope"),
     Environments = require("./config/environments"),
     blankScriptAST = require("../conf/blank-script.json"),
@@ -522,7 +521,7 @@ module.exports = (function() {
         scopeMap = null,
         scopeManager = null,
         currentFilename = null,
-        controller = null,
+        traverser = null,
         reportingConfig = [],
         sourceCode = null;
 
@@ -640,7 +639,7 @@ module.exports = (function() {
         currentScopes = null;
         scopeMap = null;
         scopeManager = null;
-        controller = null;
+        traverser = null;
         reportingConfig = [];
         sourceCode = null;
     };
@@ -786,7 +785,7 @@ module.exports = (function() {
 
             // save config so rules can access as necessary
             currentConfig = config;
-            controller = new estraverse.Controller();
+            traverser = new Traverser();
 
             ecmaFeatures = currentConfig.parserOptions.ecmaFeatures || {};
             ecmaVersion = currentConfig.parserOptions.ecmaVersion || 5;
@@ -798,8 +797,7 @@ module.exports = (function() {
                 impliedStrict: ecmaFeatures.impliedStrict,
                 ecmaVersion: ecmaVersion,
                 sourceType: currentConfig.parserOptions.sourceType || "script",
-                childVisitorKeys: espree.VisitorKeys,
-                fallback: "none"
+                fallback: Traverser.getKeys
             });
             currentScopes = scopeManager.scopes;
 
@@ -839,15 +837,14 @@ module.exports = (function() {
              * an event is fired. This allows any listeners to automatically be informed
              * that this type of node has been found and react accordingly.
              */
-            controller.traverse(ast, {
+            traverser.traverse(ast, {
                 enter: function(node, parent) {
                     node.parent = parent;
                     eventGenerator.enterNode(node);
                 },
                 leave: function(node) {
                     eventGenerator.leaveNode(node);
-                },
-                keys: espree.VisitorKeys
+                }
             });
         }
 
@@ -976,7 +973,7 @@ module.exports = (function() {
      * @returns {ASTNode[]} Array of objects representing ancestors.
      */
     api.getAncestors = function() {
-        return controller.parents();
+        return traverser.parents();
     };
 
     /**
@@ -984,14 +981,14 @@ module.exports = (function() {
      * @returns {Object} An object representing the current node's scope.
      */
     api.getScope = function() {
-        var parents = controller.parents(),
+        var parents = traverser.parents(),
             scope = currentScopes[0];
 
         // Don't do this for Program nodes - they have no parents
         if (parents.length) {
 
             // if current node introduces a scope, add it to the list
-            var current = controller.current();
+            var current = traverser.current();
             if (currentConfig.parserOptions.ecmaVersion >= 6) {
                 if (["BlockStatement", "SwitchStatement", "CatchClause", "FunctionDeclaration", "FunctionExpression", "ArrowFunctionExpression"].indexOf(current.type) >= 0) {
                     parents.push(current);