--- a/lib/rules/no-unmodified-loop-condition.js
+++ b/lib/rules/no-unmodified-loop-condition.js
@@ -12,8 +12,7 @@
 //------------------------------------------------------------------------------
 
 var Map = require("es6-map"),
-    espree = require("espree"),
-    estraverse = require("estraverse"),
+    Traverser = require("../util/traverser"),
     astUtils = require("../ast-utils");
 
 //------------------------------------------------------------------------------
@@ -116,9 +115,10 @@ var isInLoop = {
  * @returns {boolean} `true` if the node is dynamic.
  */
 function hasDynamicExpressions(root) {
-    var retv = false;
+    var retv = false,
+        traverser = new Traverser();
 
-    estraverse.traverse(root, {
+    traverser.traverse(root, {
         enter: function(node) {
             if (DYNAMIC_PATTERN.test(node.type)) {
                 retv = true;
@@ -126,8 +126,7 @@ function hasDynamicExpressions(root) {
             } else if (SKIP_PATTERN.test(node.type)) {
                 this.skip();
             }
-        },
-        keys: espree.VisitorKeys
+        }
     });
 
     return retv;