Not quite ready yet, but putting here for early feedback.

Still waiting on a new esrecurse release to npm.