@nzakas 2016-03-11
Looks like perf is just about even if not a bit better.

Before:

```
Single File:
  CPU Speed is 3093 with multiplier 13000000
  Performance Run #1:  4824.679365ms
  Performance Run #2:  4719.578955ms
  Performance Run #3:  4776.918791ms
  Performance Run #4:  4551.900144ms
  Performance Run #5:  4530.163554ms

  Performance budget exceeded: 4719.578955ms (limit: 4203.039120594892ms)


Multi Files (450 files):
  CPU Speed is 3093 with multiplier 39000000
  Performance Run #1:  14429.456481ms
  Performance Run #2:  14119.118036ms
  Performance Run #3:  14838.534102ms
  Performance Run #4:  40187.292232ms
  Performance Run #5:  13989.529158ms

  Performance budget exceeded: 14429.456481ms (limit: 12609.117361784674ms)
```

After:

```
Single File:
  CPU Speed is 3093 with multiplier 13000000
  Performance Run #1:  4598.792312ms
  Performance Run #2:  4618.367989ms
  Performance Run #3:  4389.600934ms
  Performance Run #4:  4566.863975ms
  Performance Run #5:  4447.044066ms

  Performance budget exceeded: 4566.863975ms (limit: 4203.039120594892ms)


Multi Files (450 files):
  CPU Speed is 3093 with multiplier 39000000
  Performance Run #1:  13736.515981ms
  Performance Run #2:  13696.480994ms
  Performance Run #3:  14025.821698ms
  Performance Run #4:  13909.789924ms
  Performance Run #5:  13746.484917ms

  Performance budget exceeded: 13746.484917ms (limit: 12609.117361784674ms)
```