@marco-c 2016-03-13
The error is still happening for us: https://github.com/mozilla/platatus/pull/374.