@ilyavolodin 2016-03-13
@marco-c This PR provides facility for `babel-eslint` to fix this error. You'll need to wait for `babel-eslint` to update.